auth/requirements.txt
grpcio~=1.19.0
psycopg2